機能をutils.jsに分割
